+++ diesel_tests/tests/insert.rs
@@ -549,14 +549,12 @@ fn batch_insert_with_returning_id_sqlite() {
 }
 
 #[diesel_test_helper::test]
-#[cfg(not(any(feature = "mysql", feature = "mariadb")))] // FIXME: Figure out how to handle tests that modify schema
 fn batch_insert_with_defaults() {
     use crate::schema::users::table as users;
     use crate::schema_dsl::*;
 
     let connection = &mut connection();
-    drop_table_cascade(connection, "users");
-    create_table(
+    create_temporary_table(
         "users",
         (
             integer("id").primary_key().auto_increment(),
@@ -594,14 +592,12 @@ fn batch_insert_with_defaults() {
 }
 
 #[diesel_test_helper::test]
-#[cfg(not(any(feature = "mysql", feature = "mariadb")))] // FIXME: Figure out how to handle tests that modify schema
 fn insert_with_defaults() {
     use crate::schema::users::table as users;
     use crate::schema_dsl::*;
 
     let connection = &mut connection();
-    drop_table_cascade(connection, "users");
-    create_table(
+    create_temporary_table(
         "users",
         (
             integer("id").primary_key().auto_increment(),
@@ -627,14 +623,12 @@ fn insert_with_defaults() {
 }
 
 #[diesel_test_helper::test]
-#[cfg(not(any(feature = "mysql", feature = "mariadb")))] // FIXME: Figure out how to handle tests that modify schema
 fn insert_in_nullable_with_non_null_default() {
     use crate::schema::users::table as users;
     use crate::schema_dsl::*;
 
     let connection = &mut connection();
-    drop_table_cascade(connection, "users");
-    create_table(
+    create_temporary_table(
         "users",
         (
             integer("id").primary_key().auto_increment(),
@@ -673,16 +667,14 @@ fn insert_in_nullable_with_non_null_default() {
 }
 
 #[diesel_test_helper::test]
-#[cfg(not(any(feature = "mysql", feature = "mariadb")))] // FIXME: Figure out how to handle tests that modify schema
 fn insert_returning_count_returns_number_of_rows_inserted() {
     use crate::schema::users::table as users;
     let connection = &mut connection();
-    drop_table_cascade(connection, "users");
     diesel::sql_query(
-        "CREATE TABLE users (
+        "CREATE TEMPORARY TABLE users (
         id SERIAL PRIMARY KEY,
-        name VARCHAR NOT NULL,
-        hair_color VARCHAR NOT NULL DEFAULT 'Green'
+        name VARCHAR(50) NOT NULL,
+        hair_color VARCHAR(10) NOT NULL DEFAULT 'Green'
     )",
     )
     .execute(connection)
@@ -724,7 +716,7 @@ fn insert_with_generated_column() {
 
     let connection = &mut connection();
     diesel::sql_query(
-        "CREATE TABLE user_with_last_names (
+        "CREATE TEMPORARY TABLE user_with_last_names (
         first_name VARCHAR NOT NULL PRIMARY KEY,
         last_name VARCHAR NOT NULL,
         full_name VARCHAR GENERATED ALWAYS AS (first_name || ' ' || last_name) STORED
@@ -828,8 +820,7 @@ fn insert_only_default_values_with_returning() {
     use crate::schema_dsl::*;
     let connection = &mut connection();
 
-    drop_table_cascade(connection, "users");
-    create_table(
+    create_temporary_table(
         "users",
         (
             integer("id").primary_key().auto_increment(),
@@ -1002,13 +993,11 @@ fn insert_optional_field_with_null() {
 }
 
 #[diesel_test_helper::test]
-#[cfg(not(any(feature = "mysql", feature = "mariadb")))]
 fn insert_optional_field_with_default() {
     use crate::schema::users::dsl::*;
     use crate::schema_dsl::*;
     let connection = &mut connection();
-    drop_table_cascade(connection, "users");
-    create_table(
+    create_temporary_table(
         "users",
         (
             integer("id").primary_key().auto_increment(),
@@ -1037,13 +1026,11 @@ fn insert_optional_field_with_default() {
 }
 
 #[diesel_test_helper::test]
-#[cfg(not(any(feature = "mysql", feature = "mariadb")))]
 fn insert_all_default_fields() {
     use crate::schema::users::dsl::*;
     use crate::schema_dsl::*;
     let connection = &mut connection();
-    drop_table_cascade(connection, "users");
-    create_table(
+    create_temporary_table(
         "users",
         (
             integer("id").primary_key().auto_increment(),